Extract the shared search reflector; keep the SSDP legs symmetric
The per-searcher session machinery (reserve a port, register a response
capture, dedup and evict) was SSDP-specific but identical to what WSD
needs. Move it into reflector/search.rs as a generic SearchReflector plus
a per-session ResponseReflector, parameterized by a Verdict classifier, a
session-window policy, the re-emit TTL, and a ReplyRewrite factory.

ReplyRewrite is the per-session reply transform: DialRewrite (SSDP's DIAL
LOCATION rewrite, now a trait impl replacing the free dial_rewrite fn) or
the NoRewrite ZST. The shared module stays DIAL-agnostic — SSDP injects the
DialRewrite factory, and the advertisement direction calls its rewrite
method directly.

SSDP's search leg (the classifier verdict, the MX window, and the reflector
construction) lives in ssdp/search.rs, mirroring ssdp/advertisement.rs;
DialRewrite stays in ssdp.rs since both legs use it. ssdp.rs is now a thin
coordinator.
