+++ src/reflector/search.rs
@@ -1,0 +1,553 @@
+//! The shared search direction for the unicast-reply discovery protocols (SSDP, and next WSD).
+//!
+//! A search (SSDP `M-SEARCH`, WSD `Probe` / `Resolve`) is reflected source → target, and each
+//! searcher's *unicast* reply (SSDP `200 OK`, WSD `ProbeMatches` / `ResolveMatches`) is routed back
+//! through a per-searcher session: a reserved ephemeral port on the target with a dedicated response
+//! capture, so a reply reaches only the searcher that asked. [`SearchReflector`] owns the sessions and
+//! reflects searches; a per-session [`ResponseReflector`] routes each reply back.
+//!
+//! Protocol specifics enter as parameters: the [`Verdict`] classifier (is this payload a search?), the
+//! session-window policy, the re-emit TTL, and a [`ReplyRewrite`] factory — SSDP injects its DIAL
+//! `LOCATION` rewrite, WSD uses the [`NoRewrite`] no-op.
+
+use std::net::{IpAddr, SocketAddr};
+use std::time::{Duration, Instant};
+
+use crate::dispatch::{CaptureKey, Filter, PacketDispatcher, PacketHandler, RegistrationKey};
+use crate::interface::{InterfaceAddresses, Ipv6Scope};
+use crate::net::mac::{MacAddr, MacSet};
+use crate::net::packet::Packet;
+use crate::net::port_reservation::PortReservation;
+use crate::reactor::Reactor;
+
+use super::{Verdict, egress_sources};
+
+/// In-flight session cap, so a burst of searchers can't exhaust ephemeral ports or registrations. At
+/// the cap a new search is dropped (no live session is evicted early).
+const MAX_SESSIONS: usize = 64;
+
+/// Transforms a reply datagram before it is routed back to the searcher. The returned slice is either
+/// `payload` verbatim or a rewrite held in the implementor's own reused scratch — a lending signature
+/// the `Fn` traits can't express, which is why this is a trait and not a closure.
+pub(crate) trait ReplyRewrite {
+    fn rewrite<'a>(
+        &'a mut self,
+        payload: &'a [u8],
+        egress: CaptureKey,
+        dispatcher: &mut PacketDispatcher,
+        reactor: &mut Reactor,
+    ) -> &'a [u8];
+}
+
+/// The identity transform: forward the reply verbatim. A ZST for protocols (WSD) whose replies are
+/// relayed unchanged.
+pub(crate) struct NoRewrite;
+
+impl ReplyRewrite for NoRewrite {
+    fn rewrite<'a>(
+        &'a mut self,
+        payload: &'a [u8],
+        _egress: CaptureKey,
+        _dispatcher: &mut PacketDispatcher,
+        _reactor: &mut Reactor,
+    ) -> &'a [u8] {
+        payload
+    }
+}
+
+/// One in-flight search. The searcher (`ip:port`) is the dedup key; `expiry` is when the session
+/// lapses; `reservation` holds the ephemeral target port the replies arrive on for the session's life
+/// (dropping it frees the port); `response_key` is the per-session response capture. A
+/// `RegistrationKey` is not a RAII guard, so eviction and rollback `unregister` it by hand.
+struct Session {
+    searcher: SocketAddr,
+    expiry: Instant,
+    reservation: PortReservation,
+    response_key: RegistrationKey,
+}
+
+/// One search session's reply path: a standalone leaf that re-emits each unicast reply — captured at
+/// the session's reserved port on the target — onto `egress` (the source), back to the single
+/// `searcher` that searched. It carries everything a reply needs, so no session lookup is required: the
+/// reply goes to the searcher's captured frame MAC (no ARP/ND) and is sourced from the responding
+/// device's own reply port. [`SearchReflector`] creates one per session and drops it on expiry.
+struct ResponseReflector {
+    searcher: SocketAddr,
+    searcher_mac: MacAddr,
+    egress: CaptureKey,
+    /// Protocol label for logs, e.g. `"SSDP"`.
+    name: &'static str,
+    ttl: u8,
+    reply: Box<dyn ReplyRewrite>,
+}
+
+impl PacketHandler for ResponseReflector {
+    fn on_packet(
+        &mut self,
+        packet: &Packet,
+        dispatcher: &mut PacketDispatcher,
+        reactor: &mut Reactor,
+    ) {
+        // The dispatcher's filter already pinned this capture to the reserved port, so every packet
+        // here is a unicast reply for this searcher — nothing to classify. A family the source can't
+        // currently send is a quiet drop (transient address loss).
+        if !egress_sources(dispatcher, self.egress, self.searcher) {
+            log::debug!(
+                "{}: egress has no source for searcher {} yet; dropping response from {}",
+                self.name,
+                self.searcher,
+                packet.source
+            );
+            return;
+        }
+        let payload = self
+            .reply
+            .rewrite(packet.payload, self.egress, dispatcher, reactor);
+        match dispatcher.send_udp(
+            self.egress,
+            self.searcher,
+            self.searcher_mac,
+            packet.source.port(),
+            self.ttl,
+            payload,
+        ) {
+            Ok(()) => log::debug!(
+                "reflected {} response from {} to searcher {}",
+                self.name,
+                packet.source,
+                self.searcher
+            ),
+            Err(e) => log::warn!(
+                "{}: cannot reflect response to searcher {}: {e}",
+                self.name,
+                self.searcher
+            ),
+        }
+    }
+}
+
+/// Reflects searches source → target and routes each unicast reply back to its searcher. Registered
+/// per group on the source and owns the sessions for searches to that group. On a search it dedups
+/// against live sessions (a retransmit refreshes the window and re-reflects from the same reserved
+/// port), else opens a session — reserve an ephemeral port on the target, register a
+/// [`ResponseReflector`] for its replies — and reflects the search from that port. The deadline timer
+/// sweeps expired sessions.
+///
+/// Protocol-specific behaviour is injected: `classify` gates the ingress ([`Verdict::Reflect`] = a
+/// search to handle, [`Verdict::Skip`] = the other direction, [`Verdict::Junk`] = log and drop);
+/// `window` is the per-search session lifetime; `make_reply` mints the per-session reply transform.
+pub(crate) struct SearchReflector {
+    /// The source capture: this reflector's ingress, and the egress its responses leave reply on.
+    source: CaptureKey,
+    /// The target capture: where the search is re-emitted and the replies are captured.
+    target: CaptureKey,
+    /// The target interface's index, for the IPv6 link-local reserved-port bind.
+    target_ifindex: u32,
+    /// The configured device allow-set, scoping the response capture as the announcement direction is.
+    device_macs: Option<MacSet>,
+    /// Protocol label for logs, e.g. `"SSDP"`.
+    name: &'static str,
+    /// The TTL each reflected search and reply is re-emitted at.
+    ttl: u8,
+    /// The ingress gate: is this payload a search for this direction?
+    classify: fn(&[u8]) -> Verdict,
+    /// A search's session lifetime (e.g. SSDP's MX window + grace; a fixed value for WSD).
+    window: fn(&[u8]) -> Duration,
+    /// Mints a fresh reply transform per session (its own scratch, for a rewriting protocol).
+    make_reply: Box<dyn Fn() -> Box<dyn ReplyRewrite>>,
+    sessions: Vec<Session>,
+}
+
+impl SearchReflector {
+    #[allow(clippy::too_many_arguments)] // each is a distinct protocol parameter; grouping them buys nothing
+    pub(crate) fn new(
+        source: CaptureKey,
+        target: CaptureKey,
+        target_ifindex: u32,
+        device_macs: Option<MacSet>,
+        name: &'static str,
+        ttl: u8,
+        classify: fn(&[u8]) -> Verdict,
+        window: fn(&[u8]) -> Duration,
+        make_reply: Box<dyn Fn() -> Box<dyn ReplyRewrite>>,
+    ) -> Self {
+        Self {
+            source,
+            target,
+            target_ifindex,
+            device_macs,
+            name,
+            ttl,
+            classify,
+            window,
+            make_reply,
+            sessions: Vec::new(),
+        }
+    }
+
+    /// Open a session for a new searcher: reserve an ephemeral port on the target's own address of the
+    /// search's family and register the reply capture there — before the caller reflects, so a fast
+    /// responder can't beat the capture. `None` (logged) if the cap is hit, the frame carries no
+    /// source MAC to reply to, the target has no address of that family, or the reservation fails.
+    fn make_session(
+        &self,
+        packet: &Packet,
+        dispatcher: &mut PacketDispatcher,
+        expiry: Instant,
+    ) -> Option<Session> {
+        if self.sessions.len() >= MAX_SESSIONS {
+            log::warn!(
+                "{}: dropping search from {}: {MAX_SESSIONS} sessions in flight (cap)",
+                self.name,
+                packet.source
+            );
+            return None;
+        }
+        let Some(searcher_mac) = packet.src_mac else {
+            log::warn!(
+                "{}: cannot reflect search from {}: frame has no source MAC to reply to",
+                self.name,
+                packet.source
+            );
+            return None;
+        };
+        // The replies come to the address the reflected search is sourced from, at the reserved port —
+        // so this must be the same scope-matched pick `build_udp` makes for `packet.dest`, or the
+        // device replies to a source the response capture below isn't watching.
+        let our_addr = match packet.dest.ip() {
+            IpAddr::V4(_) => dispatcher
+                .egress_addrs(self.target)
+                .and_then(InterfaceAddresses::v4)
+                .map(IpAddr::V4),
+            IpAddr::V6(dst6) => dispatcher
+                .egress_addrs(self.target)
+                .and_then(|a| a.v6(Ipv6Scope::of(dst6)))
+                .map(IpAddr::V6),
+        };
+        let Some(our_addr) = our_addr else {
+            log::warn!(
+                "{}: cannot reflect search from {}: target has no source address for {}",
+                self.name,
+                packet.source,
+                packet.dest.ip()
+            );
+            return None;
+        };
+        let reservation = match PortReservation::create(our_addr, self.target_ifindex) {
+            Ok(reservation) => reservation,
+            Err(e) => {
+                log::warn!(
+                    "{}: port reservation for searcher {} failed: {e}",
+                    self.name,
+                    packet.source
+                );
+                return None;
+            }
+        };
+        // Register before the reflect so a fast responder's reply is captured, not ICMP-rejected.
+        let response_key = dispatcher.register(
+            self.target,
+            Filter {
+                dst_ip: Some(our_addr.into()),
+                dst_port: Some(reservation.port().into()),
+                src_mac: self.device_macs.clone(),
+                ..Filter::default()
+            },
+            Box::new(ResponseReflector {
+                searcher: packet.source,
+                searcher_mac,
+                egress: self.source,
+                name: self.name,
+                ttl: self.ttl,
+                reply: (self.make_reply)(),
+            }),
+        );
+        Some(Session {
+            searcher: packet.source,
+            expiry,
+            reservation,
+            response_key,
+        })
+    }
+}
+
+impl PacketHandler for SearchReflector {
+    fn on_packet(
+        &mut self,
+        packet: &Packet,
+        dispatcher: &mut PacketDispatcher,
+        _reactor: &mut Reactor,
+    ) {
+        match (self.classify)(packet.payload) {
+            Verdict::Reflect => {}
+            // A message for the other direction (an announcement) flows through that reflector.
+            Verdict::Skip => return,
+            Verdict::Junk => {
+                log::debug!(
+                    "{}: dropping unrecognized payload ({} B) on the search path from {}",
+                    self.name,
+                    packet.payload.len(),
+                    packet.source
+                );
+                return;
+            }
+        }
+        let expiry = Instant::now() + (self.window)(packet.payload);
+
+        // A retransmit from a known searcher reuses its session: refresh the window and re-reflect from
+        // the same reserved port. A new searcher opens a fresh session.
+        if let Some(session) = self
+            .sessions
+            .iter_mut()
+            .find(|s| s.searcher == packet.source)
+        {
+            let port = session.reservation.port();
+            match dispatcher.send_udp_group(
+                self.target,
+                packet.dest,
+                port,
+                self.ttl,
+                packet.payload,
+            ) {
+                Ok(()) => {
+                    session.expiry = expiry;
+                    log::debug!(
+                        "re-reflected {} search from {} to {} on reserved port {port}",
+                        self.name,
+                        packet.source,
+                        packet.dest
+                    );
+                }
+                Err(e) => log::warn!(
+                    "{}: cannot reflect search from {} to {}: {e}",
+                    self.name,
+                    packet.source,
+                    packet.dest
+                ),
+            }
+            return;
+        }
+
+        let Some(session) = self.make_session(packet, dispatcher, expiry) else {
+            return; // make_session logged the cause
+        };
+        let port = session.reservation.port();
+        match dispatcher.send_udp_group(self.target, packet.dest, port, self.ttl, packet.payload) {
+            Ok(()) => {
+                self.sessions.push(session);
+                log::debug!(
+                    "reflected {} search from {} to {} on reserved port {port}; opened a session, {} active",
+                    self.name,
+                    packet.source,
+                    packet.dest,
+                    self.sessions.len()
+                );
+            }
+            Err(e) => {
+                // Roll back the response capture just registered; the reservation drops with `session`.
+                log::warn!(
+                    "{}: cannot reflect search from {} to {}: {e}",
+                    self.name,
+                    packet.source,
+                    packet.dest
+                );
+                dispatcher.unregister(session.response_key);
+            }
+        }
+    }
+
+    fn next_deadline(&self) -> Option<Instant> {
+        self.sessions.iter().map(|s| s.expiry).min()
+    }
+
+    fn on_deadline(
+        &mut self,
+        now: Instant,
+        dispatcher: &mut PacketDispatcher,
+        _reactor: &mut Reactor,
+    ) {
+        self.sessions.retain(|session| {
+            if session.expiry <= now {
+                dispatcher.unregister(session.response_key);
+                log::debug!(
+                    "evicted {} session for searcher {} on reserved port {}",
+                    self.name,
+                    session.searcher,
+                    session.reservation.port()
+                );
+                false
+            } else {
+                true
+            }
+        });
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::net::Ipv4Addr;
+
+    use super::*;
+
+    const TEST_TTL: u8 = 2;
+
+    /// A trivial ingress gate: every payload is a search. The session bookkeeping under test is
+    /// independent of how a protocol classifies.
+    fn always_reflect(_: &[u8]) -> Verdict {
+        Verdict::Reflect
+    }
+
+    /// A fixed session window, standing in for a protocol's window policy.
+    fn fixed_window(_: &[u8]) -> Duration {
+        Duration::from_secs(2)
+    }
+
+    fn test_reflector() -> SearchReflector {
+        SearchReflector::new(
+            CaptureKey::from_u64(1),
+            CaptureKey::from_u64(0),
+            0,
+            None,
+            "TEST",
+            TEST_TTL,
+            always_reflect,
+            fixed_window,
+            Box::new(|| Box::new(NoRewrite) as Box<dyn ReplyRewrite>),
+        )
+    }
+
+    /// Push a session for `searcher` onto `reflector`: a real loopback port reservation plus a
+    /// registered response capture, so eviction has a registration to tear down. (`PortReservation`
+    /// binds a socket directly, so no capture / `CAP_NET_RAW` is needed.)
+    fn push_session(
+        reflector: &mut SearchReflector,
+        dispatcher: &mut PacketDispatcher,
+        searcher: &str,
+        expiry: Instant,
+    ) {
+        let searcher: SocketAddr = searcher.parse().unwrap();
+        let (target, source) = (reflector.target, reflector.source);
+        let reservation = PortReservation::create(IpAddr::V4(Ipv4Addr::LOCALHOST), 0)
+            .expect("reserve a loopback port");
+        let response_key = dispatcher.register(
+            target,
+            Filter::default(),
+            Box::new(ResponseReflector {
+                searcher,
+                searcher_mac: MacAddr::from([0; 6]),
+                egress: source,
+                name: "TEST",
+                ttl: TEST_TTL,
+                reply: Box::new(NoRewrite),
+            }),
+        );
+        reflector.sessions.push(Session {
+            searcher,
+            expiry,
+            reservation,
+            response_key,
+        });
+    }
+
+    #[test]
+    fn next_deadline_is_the_soonest_session_expiry() {
+        let mut dispatcher = PacketDispatcher::new();
+        let mut reflector = test_reflector();
+        assert_eq!(
+            reflector.next_deadline(),
+            None,
+            "no sessions means no timer"
+        );
+        let base = Instant::now();
+        push_session(
+            &mut reflector,
+            &mut dispatcher,
+            "10.0.0.1:5",
+            base + Duration::from_secs(5),
+        );
+        push_session(
+            &mut reflector,
+            &mut dispatcher,
+            "10.0.0.2:5",
+            base + Duration::from_secs(2),
+        );
+        assert_eq!(
+            reflector.next_deadline(),
+            Some(base + Duration::from_secs(2))
+        );
+    }
+
+    #[test]
+    fn on_deadline_evicts_expired_sessions_and_unregisters_their_captures() {
+        let mut dispatcher = PacketDispatcher::new();
+        let mut reactor = Reactor::new().unwrap();
+        let mut reflector = test_reflector();
+        let base = Instant::now();
+        push_session(&mut reflector, &mut dispatcher, "10.0.0.1:5", base); // already due
+        push_session(
+            &mut reflector,
+            &mut dispatcher,
+            "10.0.0.2:5",
+            base + Duration::from_secs(10),
+        ); // live
+        assert_eq!(dispatcher.registration_count(), 2);
+
+        reflector.on_deadline(base + Duration::from_secs(1), &mut dispatcher, &mut reactor);
+
+        assert_eq!(
+            reflector.sessions.len(),
+            1,
+            "the expired session is dropped"
+        );
+        assert_eq!(
+            reflector.sessions[0].searcher,
+            "10.0.0.2:5".parse::<SocketAddr>().unwrap()
+        );
+        assert_eq!(
+            dispatcher.registration_count(),
+            1,
+            "its response capture is unregistered with it"
+        );
+        assert_eq!(
+            reflector.next_deadline(),
+            Some(base + Duration::from_secs(10))
+        );
+    }
+
+    #[test]
+    fn a_retransmit_reuses_its_session_and_refreshes_the_window() {
+        let mut dispatcher = PacketDispatcher::new();
+        let mut reactor = Reactor::new().unwrap();
+        // A synthetic target: send_udp_group on an unknown egress drops the datagram and returns Ok,
+        // so the re-reflect "succeeds" with no real capture — this exercises only the bookkeeping.
+        let mut reflector = test_reflector();
+        let base = Instant::now();
+        push_session(&mut reflector, &mut dispatcher, "10.0.0.7:50000", base);
+        assert_eq!(dispatcher.registration_count(), 1);
+
+        let packet = Packet {
+            source: "10.0.0.7:50000".parse().unwrap(),
+            dest: "239.255.255.250:1900".parse().unwrap(),
+            ttl: TEST_TTL,
+            dst_mac: None,
+            src_mac: Some(MacAddr::from([0x02, 0, 0, 0, 0, 1])),
+            payload: b"a search",
+        };
+        reflector.on_packet(&packet, &mut dispatcher, &mut reactor);
+
+        assert_eq!(
+            reflector.sessions.len(),
+            1,
+            "a retransmit reuses its session, not a new one"
+        );
+        assert_eq!(
+            dispatcher.registration_count(),
+            1,
+            "no second response capture is registered"
+        );
+        assert!(
+            reflector.sessions[0].expiry > base,
+            "the session's window is refreshed"
+        );
+    }
+}
